@@ -16,8 +16,7 @@ Gem::Specification.new do |s|
   s.licenses    = 'MIT'
   s.require_paths = %w[lib]
 
-  s.add_dependency('rails', '>= 5.1')
-  s.add_development_dependency('mocha')
+  s.add_dependency('rails', '~> 7.0')
   s.add_development_dependency('faker')
   s.add_development_dependency('rake')
   s.add_development_dependency('rubocop')
