@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class AddTypesToQuestionsAndOptions < ActiveRecord::Migration
+class AddTypesToQuestionsAndOptions < ActiveRecord::Migration[7.0]
   def change
     # Survey Questions table
     add_column :survey_questions, :questions_type_id, :integer
