@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateLessons < ActiveRecord::Migration
+class CreateLessons < ActiveRecord::Migration[7.0]
   def change
     create_table :lessons do |t|
       t.string :name
