@@ -2,16 +2,12 @@
 
 class Survey::Question < ActiveRecord::Base
   self.table_name = 'survey_questions'
+
   # relations
   has_many   :options
   has_many   :predefined_values
   has_many   :answers
   belongs_to :section
-
-  # rails 3 attr_accessible support
-  if Rails::VERSION::MAJOR < 4
-    attr_accessible :options_attributes, :predefined_values_attributes, :text, :section_id, :head_number, :description, :locale_text, :locale_head_number, :locale_description, :questions_type_id
-  end
 
   accepts_nested_attributes_for :options,
                                 reject_if: ->(a) { a[:options_type_id].blank? },
@@ -22,9 +18,15 @@ class Survey::Question < ActiveRecord::Base
                                 allow_destroy: true
 
   # validations
-  validates :text, presence: true, allow_blank: false
-  validates :questions_type_id, presence: true
-  validates :questions_type_id, inclusion: { in: Survey::QuestionsType.questions_type_ids, unless: proc { |q| q.questions_type_id.blank? } }
+  validates :text,
+            presence: true,
+            allow_blank: false
+
+  validates :questions_type_id,
+            presence: true
+
+  validates :questions_type_id,
+            inclusion: { in: Survey::QuestionsType.questions_type_ids, unless: proc { |q| q.questions_type_id.blank? } }
 
   scope :mandatory_only, -> { where(mandatory: true) }
 
